Tratando erros de integridade e listando anúncios cadastrados.

+++ app/Http/Controllers/Admin/CategoriaController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
+use App\Models\Anuncio;
 use App\Models\Categoria;
 
 class CategoriaController extends Controller
@@ -54,6 +55,17 @@ class CategoriaController extends Controller
 
     public function remover(Request $request, $id)
     {
+        if (Anuncio::where('categoria_id', '=', $id)->count() > 0) {
+            $msg = "Não é possível remover uma categoria que possua anúncios vinculados:";
+            $anuncios = Anuncio::where('categoria_id', '=', $id)->get();
+            foreach ($anuncios as $anuncio) {
+                $msg .= " id:" . $anuncio->id;
+            }
+            $request->session()->flash('mensagem',
+                ['msg'=>$msg, 'class'=>'red white-text']);
+            return redirect()->route('admin.categorias');
+        }
+
         Categoria::find($id)->delete();
 
         $request->session()->flash('mensagem',
